more auto

+++ dad-is-great-in-C/ognfunc.hpp
@@ -9,4 +9,4 @@ unsigned long long reverse(unsigned long long n);
 unsigned long long numsys(unsigned long long n, unsigned char k);
 unsigned short numcount(unsigned long long n);
 std::string numsys16(unsigned long long n, unsigned char k);
-double binpow(double b, unsigned short e);
+long double binpow(long double b, unsigned short e);
